Update login-form.jsx
We do not need to fire a Redux action every time the user clicks a button and cause React to re-render. This is inefficient because the UI does not change. If we keep the user input in the local component state, we can avoid not needed re-renders and changes to the global state. The global state does not need to know about incomplete changes. For example: A user is creating an account on a Signup page. We do not need to make an API call every time the user types a letter for their name or phone number. We only need to do that when the user has finished typing and is ready to submit the entire form.